Separate RapCloud & Settings
